fix: cast severity count for CompileStatic

+++ src/main/groovy/com/boghus/codereview/review/ReviewResult.groovy
@@ -12,7 +12,7 @@ class ReviewResult {
     }
 
     int count(String severity) {
-        findings.count { it.severity.equalsIgnoreCase(severity) }
+        (int) findings.count { it.severity.equalsIgnoreCase(severity) }
     }
 
     boolean hasCriticalOrHigh() {
